+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ﻿### What's new?
 
+* __If you run the executable without arguments (e.g: from the file explorer), you'll be guided on providing
+the arguments it needs via prompts.__
 * The program now checks for duplicate collections and will ask you if you want to continue if it finds one.
 You can also disable this behaviour by passing the `--skip-duplicate-check` flag.
 * Backups now include the date in the filename to not overwrite the old ones.